[REACTOS] Fix x64 build - SIZE_T -> ULONG in KsecGatherEntropyData - Add missing ZwQueryInformationProcess() prototype for x64 build of btrfs - Fix ml.exe path for VS 2017 Community Edition - Add missing Handle32ToHandle to basetsd.h

+++ drivers/crypto/ksecdd/random.c
@@ -96,7 +96,7 @@ KsecGatherEntropyData(
     PTEB Teb;
     PPEB Peb;
     PWSTR String;
-    SIZE_T ReturnLength;
+    ULONG ReturnLength;
     NTSTATUS Status;
 
     /* Query some generic values */
